chore(shelly): refactor categorizeMessage to serve as an example for other ai actions
