+++ extensions/shelly/actions/categorizeMessage/lib/categorizeMessageWithLLM/categorizeMessageWithLLM.ts
@@ -1,7 +1,24 @@
-import { parser, systemPrompt } from './constants'
+import { parser } from './parser'
+import { systemPrompt } from './prompt'
 import { type ChatOpenAI } from '@langchain/openai'
+import { type AIActionMetadata } from '../../../../../../src/lib/llm/openai/types'
 
-// TODO: remove console logs eventually
+/**
+ * Uses LLM to categorize a message into predefined categories.
+ * The function follows these steps:
+ * 1. Formats prompt with available categories
+ * 2. Runs LLM chain with structured output parsing
+ * 3. Validates and processes the result
+ * 
+ * @example
+ * const result = await categorizeMessageWithLLM({
+ *   model,
+ *   message: "I need to schedule an appointment",
+ *   categories: ["Scheduling", "Medical Question"],
+ *   metadata: { ... }
+ * })
+ * // Returns: { category: "Scheduling", explanation: "..." }
+ */
 export const categorizeMessageWithLLM = async ({
   model,
   message,
@@ -11,11 +28,7 @@ export const categorizeMessageWithLLM = async ({
   model: ChatOpenAI
   message: string
   categories: string[]
-  metadata: {
-    care_flow_definition_id: string
-    care_flow_id: string
-    activity_id: string
-  }
+  metadata: AIActionMetadata  // Using dedicated metadata type
 }): Promise<{ category: string; explanation: string }> => {
   const prompt = await systemPrompt.format({
     categories: categories.concat('None').join(', '),
@@ -32,7 +45,6 @@ export const categorizeMessageWithLLM = async ({
       { metadata, runName: 'ShellyCategorizeMessage' }
     )
   } catch (error) {
-    console.error('Error invoking the chain:', error)
     throw new Error(
       'Failed to categorize the message due to an internal error.'
     )
